feat(gar): add gateway weights to to regsitry
We do all the work to compute these, but do not store them anywhere. Rather than doing more work on reads (and thus pay more gas to a CU) to compute them, we can store the weights at the beginning of an epoch and surface them in the network portal. It is important to note that weights will only change at the start of the next epoch.